List Product Form Input Validation
The form now requires quantity to be non-empty, the save to draft button is enabled only when new changes are made to the form, and the post listing button is enabled only when new changes are made and all field inputs are valid. Variant options are also required to have at least one option. Removed location from schema.